+++ packages/eas-cli/src/commands/channel/rollout.ts
@@ -152,11 +152,6 @@ export default class ChannelRollout extends EasCommand {
     if (argsAndFlags.nonInteractive) {
       await new NonInteractiveRollout(argsAndFlags).runAsync(ctx);
     } else {
-      Log.addNewLineIfNone();
-      Log.warn(
-        `✨ This command is in Developer Preview and has not been released to production yet. Website support is coming soon.`
-      );
-      Log.addNewLineIfNone();
       await new RolloutMainMenu(argsAndFlags).runAsync(ctx);
     }
   }